pyproject.toml
[project]
name = "dj-rest-kit"
version = "0.0.8"
authors = [{ name="Aman Bothra", email="amanbothra1777@gmail.com"}]
maintainers = [{ name="Prashant Gabu", email="prashantgabu.earn@gmail.com"}]
description = "dj-rest-kit is a reusable Django application "
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "Django>=3.2",
    "django-filter",
    "django-model-utils",
    "djangorestframework>3.3",
    "setuptools",
    "pytz",
    "python-dateutil",
    "django-json-widget==1.1.1",
]

[project.urls]
Homepage = "https://github.com/prashantgabu-e/dj-rest-kit"
Issues = "https://github.com/prashantgabu-e/dj-rest-kit/issues"

[tool.flit.module]
name = "src"